chore: add new PDF and KMZ links for intake5
