--- app/data/zoneKmzLinksIntake4.ts
@@ -1,60 +1,0 @@
-const kmzAndPdfLinks = {
-  1: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/16958a8e-108c-4215-a907-2e33e921b0a1/download/areas-of-interest_intake_4_zone1_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/2ad10eb3-3dcc-40d7-b92e-ad7977b6ebf1/download/zone-1-areas-of-interest.kmz',
-  },
-  2: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/97912049-7f67-44ea-b329-92909d93a45e/download/areas-of-interest_intake_4_zone2_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/34f37e98-9465-4ca3-86f2-3dedfbd1698a/download/zone-2-areas-of-interest.kmz',
-  },
-  3: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/cc1bacc3-21f3-4af5-b40a-762f36becde4/download/areas-of-interest_intake_4_zone3_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/66984771-50d2-4b18-aa21-b7207d803c45/download/zone-3-areas-of-interest.kmz',
-  },
-  4: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/f7270db2-2dd1-4811-a451-2d025c7804fc/download/areas-of-interest_intake_4_zone4_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/6d87a908-3536-47e8-b910-8f81561e29e4/download/zone-4-areas-of-interest.kml',
-  },
-  5: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/d416e327-d39b-4ae2-a919-a153dc2c606f/download/areas-of-interest_intake_4_zone5_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/f5fdb9a7-9953-4bf1-a397-968be7fd4a72/download/zone-5-areas-of-interest.kml',
-  },
-  6: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/e6b239d8-484e-49ed-a348-0c83da997a49/download/areas-of-interest_intake_4_zone6_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/dfe008ca-8e58-448d-9d8a-bef9b51e3723/download/zone-6-areas-of-interest.kmz',
-  },
-  7: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/a664a4ec-5a43-4471-ba4f-b31913042cb9/download/areas-of-interest_intake_4_zone7_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/cb7daecd-a7f4-4cbe-a76a-f144b3cdc6a2/download/zone-7-areas-of-interest.kml',
-  },
-  8: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/46cc6e79-02ae-411e-91a2-abc739ff2f80/download/areas-of-interest_intake_4_zone8_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/475f65e1-d7bd-49df-b824-73c45ffcebd9/download/zone-8-areas-of-interest.kml',
-  },
-  9: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/9d978cf5-8cbd-45fb-8d26-1f32ee417270/download/areas-of-interest_intake_4_zone9_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/290bf889-65b4-497a-b07f-8f874453845b/download/zone-9-areas-of-interest.kml',
-  },
-  10: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/81d9acf5-4aed-4d3c-a2c5-b7b270b1b135/download/areas-of-interest_intake_4_zone10_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/cf95b3ca-a1c9-40e9-91b6-d2f29c96f7b1/download/zone-10-areas-of-interest.kml',
-  },
-  11: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/76084e3b-6c17-4af6-aa9b-365f3e38c949/download/areas-of-interest_intake_4_zone11_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/ffbb7d24-e663-4543-8b9f-cb9f9eba232f/download/zone-11-areas-of-interest.kml',
-  },
-  12: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/3dc626ec-5330-467a-b282-b7e2aa956368/download/areas-of-interest_intake_4_zone12_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/49781801-65d1-4530-97c4-cb825b986a38/download/zone-12-areas-of-interest.kml',
-  },
-  13: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/367c6aec-204e-4a49-8ff7-94de1c740a2d/download/areas-of-interest_intake_4_zone13_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/6fc0b687-4c43-44c6-923b-1a953af87c99/download/zone-13-areas-of-interest.kml',
-  },
-  14: {
-    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/897ad72e-3dc2-4993-935b-94ea9a1e1b3e/download/areas-of-interest_intake_4_zone14_17x22.pdf',
-    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/261be538-f19d-49b6-bf3a-eddde48831e8/download/zone-14.kml',
-  },
-};
-
-export default kmzAndPdfLinks;
