+++ app/data/zoneKmzLinksIntake4And5.ts
@@ -1,0 +1,60 @@
+const kmzAndPdfLinks = {
+  1: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/84bffabb-4165-4739-a1df-5c316a473e01/download/intake_4_and_5_17x22_zone-1.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/2ad10eb3-3dcc-40d7-b92e-ad7977b6ebf1/download/zone-1-areas-of-interest.kml',
+  },
+  2: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/419f0ead-70da-48b9-9eaa-ce27b3717220/download/intake_4_and_5_17x22_zone-2.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/34f37e98-9465-4ca3-86f2-3dedfbd1698a/download/zone-2-areas-of-interest.kml',
+  },
+  3: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/e8d2657b-2025-4614-ad3d-2f9dcb4a0125/download/intake_4_and_5_17x22_zone-3.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/66984771-50d2-4b18-aa21-b7207d803c45/download/zone-3-areas-of-interest.kml',
+  },
+  4: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/23aafee4-c3b7-4b1b-95d3-32e0ec3e37a0/download/intake_4_and_5_17x22_zone-4.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/6d87a908-3536-47e8-b910-8f81561e29e4/download/zone-4-areas-of-interest.kml',
+  },
+  5: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/6486aad3-fc1c-476b-82f9-53cfe10934e1/download/intake_4_and_5_17x22_zone-5.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/f5fdb9a7-9953-4bf1-a397-968be7fd4a72/download/zone-5-areas-of-interest.kml',
+  },
+  6: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/6dd2c400-2e18-46ec-8383-4fb0d2013018/download/intake_4_and_5_17x22_zone-6.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/dfe008ca-8e58-448d-9d8a-bef9b51e3723/download/zone-6-areas-of-interest.kml',
+  },
+  7: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/e86e22c4-cfa2-47e6-8e90-fd4756478711/download/intake_4_and_5_17x22_zone-7.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/cb7daecd-a7f4-4cbe-a76a-f144b3cdc6a2/download/zone-7-areas-of-interest.kml',
+  },
+  8: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/725c0869-3c74-4eca-817a-42cd5625ae6a/download/intake_4_and_5_17x22_zone-8.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/475f65e1-d7bd-49df-b824-73c45ffcebd9/download/zone-8-areas-of-interest.kml',
+  },
+  9: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/e073c1ba-14e8-4b03-8e52-6f99049021cf/download/intake_4_and_5_17x22_zone-9.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/290bf889-65b4-497a-b07f-8f874453845b/download/zone-9-areas-of-interest.kml',
+  },
+  10: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/c99fec4a-161d-483a-8910-4cb50faaf68a/download/intake_4_and_5_17x22_zone-10.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/cf95b3ca-a1c9-40e9-91b6-d2f29c96f7b1/download/zone-10-areas-of-interest.kml',
+  },
+  11: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/c6a492e8-bdc1-4a21-8e98-592b34bcaa65/download/intake_4_and_5_17x22_zone-11.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/ffbb7d24-e663-4543-8b9f-cb9f9eba232f/download/zone-11-areas-of-interest.kml',
+  },
+  12: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/f9704e20-bf93-491e-80ea-266ff6effd59/download/intake_4_and_5_17x22_zone-12.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/49781801-65d1-4530-97c4-cb825b986a38/download/zone-12-areas-of-interest.kml',
+  },
+  13: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/f1245aed-04b6-4b50-a098-3a5beec9b954/download/intake_4_and_5_17x22_zone-13.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/6fc0b687-4c43-44c6-923b-1a953af87c99/download/zone-13-areas-of-interest.kml',
+  },
+  14: {
+    pdf: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/26e0b63e-e898-4075-84c4-8bf57b65412f/download/intake_4_and_5_17x22_zone-14.pdf',
+    kmz: 'https://catalogue.data.gov.bc.ca/dataset/332a5676-9171-4c50-add3-5c590fd7b985/resource/261be538-f19d-49b6-bf3a-eddde48831e8/download/zone-14.kml',
+  },
+};
+
+export default kmzAndPdfLinks;
